Matching Questions Finished
This was a gigantic pain, due to a bug in the location that I positioned
the XML tags.  It works fine now, except for a bug that if an answer is
exactly the same, it will create two separate options instead of linking
to the one.  Will work on it in a bit.